requirements.txt
annotated-types==0.6.0
anyio==4.3.0
black==24.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.2
Flask-Markdown==0.3
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-bigquery==3.18.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
openai==1.13.3
packaging==23.2
pathspec==0.12.1
pinatapy-vourhey==0.2.0
platformdirs==4.2.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.1
